Make the class reference table sortable (#28)
The main table's headers are now clickable to sort columns in ascending or descending order.
String columns are sorted alphabetically, description columns are sorted by MISSING vs. OK, and numeric columns are sorted by relative percentages.
A custom JS sorting implementation is used, both to avoid a large dependency and to handle the unique format of these table entries.
The cursor when hovering over table headers indicates that they can be interacted with.